requirements.txt
tensorflow-rocm

sklearn
matplotlib
pandas
scipy
numpy
h5py
Pillow

opencv-python

tqdm